% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacktrace-class.R
\name{as.stacktrace}
\alias{as.stacktrace}
\alias{print.stacktrace}
\alias{as.character.stacktrace}
\title{Add a stack trace to an error or warning.}
\usage{
as.stacktrace(condition, start_key = NULL, end_key = NULL)

\method{print}{stacktrace}(x, ...)

\method{as.character}{stacktrace}(x)
}
\arguments{
\item{condition}{Any \code{list} or S3 object of \code{\link{mode} list} to
which a stack trace should be appended; anticipated to be a \code{warning}
or \code{error} object (or anything else that inherits from the
\code{\link{condition}} class).}

\item{start_key, end_key}{a regex string that matches the text of a call in
the stack trace and marking the start (\code{start_key}) and end
(\code{end_key}) of a relevant portion of the stack call to be extracted.
The calls between the last call matching \code{start_key} and the first
call matching \code{end_key} (non_inclusive) will be included in
\code{$stack_trace}.  If \code{start_key} is NULL or doesn't match, calls
starting with the first call in the stack trace will be included  If
\code{end_key} is NULL or doesn't match, calls through the last call will
be included.  Also, the entire stack trace will be included if the last
call matching \code{start_key} occurs after the first call matching
\code{end_key}.}

\item{x}{A \code{stacktrace} object}

\item{...}{Additional parameters passed from the generic
\code{\link{print}()} function that calls \code{print.stacktrace()} (see
\code{\link{genericFunction-class}})}
}
\value{
An object of class \code{stacktrace} that inherits from the class of
  x.  If x also inherits from \code{error} or \code{warning}, result will
  have addtional class of \code{stacktraceError} or \code{stacktraceWarning},
  respectively. Result will also have an additional member in the underlying
  list, called \code{$stack_trace} containing the stack_trace (as a
  \code{character}) and as excerpted by \code{start_key} and \code{end_key}.
}
\description{
Creates a \code{stacktrace} object, typically from a \code{warning} or
\code{error} or other \code{\link{condition}} object by appending the current
stack trace or a subsection thereof as a member called \code{$stack_trace}.
}
\details{
The class \code{stacktrace} can be built atop and S3 class of
\code{\link{mode} list}, but is anticipated to inherit from the
\code{\link{condition}} class. \code{stacktrace} adds an additional member
(itself a list) called \code{$stack_trace} to the underlying list of the
\code{condition} argument.  The \code{$stack_trace} member contains results
of \code{\link{sys.calls}()}, as executed from within \code{as.stacktrace()}
and converted to \code{character}. \code{start_key} and \code{end_key}
parameters can be used to extract a relevant section from the middle of a
full stack trace.
}
