% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command.R
\name{batch_log}
\alias{batch_log}
\title{Wrapper for base::log function.}
\usage{
batch_log(x, base = exp(1))
}
\arguments{
\item{x, base}{See \code{\link{log}}.}

\item{...}{Ignored. See Details.}
}
\description{
A wrapper for \code{\link{log}} ensures that x and base are passed to log as
numeric values and includes `...` in parameter list.
}
\details{
Used as a function to test and debug the \code{\link{batch}} function.

Depending of the format of the command line arguem
}
