% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command.R
\name{is_something}
\alias{is_something}
\title{Is a value something other than NULL or FALSE?}
\usage{
is_something(x)
}
\arguments{
\item{x}{value to be tested.}
}
\value{
If x is NULL or FALSE, returns FALSE, otherwise return TRUE
}
\description{
Useful for investigating if a parameter in a command line list, which can be
missing or FALSE to represent FALSE.
}
\examples{
is_something(NULL)
is_something(FALSE)
is_something("Hello")
is_something(45)
}
