% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/command.R
\name{recursive_setNames}
\alias{recursive_setNames}
\title{Recursively convert names of a list using a function.}
\usage{
recursive_setNames(x, fun)
}
\arguments{
\item{x}{An object of \code{\ref{mode}} list (including data.frames).}

\item{fun}{A function that takes a character vector argument and returns a
character vector of the same length.}
}
\value{
The object x with environments converted to lists and names of all
nested named objects equal to fun(names(object)).
}
\description{
Replaces names of x and any named elements within x with the value with
fun(x). Operates recursively on nested lists.  Environments are converted
to named lists.
}
