% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_args.R
\name{commandArgsList}
\alias{commandArgsList}
\title{Parse command line args}
\usage{
commandArgsList(args = NULL)
}
\arguments{
\item{args}{Typlically NULL (the default), in which case \code{args} will be
derived from a call to \code{\link{commandArgs}()}.  Otherwise, a character
string of arguments that simulates the result of a call to
\code{\link{commandArgs}}; useful for debugging in interactive mode.}
}
\value{
A named list representing the parsed results of args or, if args is
  NULL, a named list representing a call to \code{\link{commandArgs}()}.
}
\description{
Converts command line arguments into a named list.
}
\details{
Attempts to convert the sometimes clunky results from
\code{\link{commandArgs}()} to a more useful named \code{list} containing
command line parameters.

If passed to \code{commandArgsList()}, \code{args} is expected to mimic the
results of a call to \code{\link{commandArgs}()}; e.g., a character vector of
arguments, by convention, in the form \code{"--<arg-name>"} or
\code{"--<arg-name>=<value>"}. However, \code{commandArgsList()} attempts to
parse the names and values of any arguments in args, regardless of format, so
long as the name and value are separated by "=".

Command line flags (arguments without values) are returned in the resulting
list with a value of TRUE.  Values assigned to arguments are returned within
the list as strings.
}
